test: triggers + refactor settings
