--- default.toml
@@ -1,22 +1,0 @@
-[app]
-hostname = "0.0.0.0"
-port = 2875
-database_url = "postgres://autopulse:autopulse@localhost:5432/autopulse"
-log_level = "info"
-
-[auth]
-username = "admin"
-password = "password"
-
-[opts]
-check_path = false
-max_retries = 5
-default_timer_wait = 60
-cleanup_days = 10
-
-[triggers.manual]
-type = "manual"
-
-[targets]
-
-[webhooks]
