+++ src/tests/utils/generate_uuid.rs
@@ -1,0 +1,10 @@
+#[cfg(test)]
+mod tests {
+    use crate::utils::generate_uuid::generate_uuid;
+
+    #[test]
+    fn test_generate_uuid() {
+        let uuid = generate_uuid();
+        assert_eq!(uuid.len(), 36);
+    }
+}
