+++ src/tests/utils/mod.rs
@@ -1,0 +1,6 @@
+pub mod check_auth;
+pub mod checksum;
+pub mod generate_uuid;
+pub mod join_path;
+pub mod rewrite;
+pub mod sify;
