+++ src/utils/checksum.rs
@@ -22,27 +22,3 @@ pub fn sha256checksum(file_path: &PathBuf) -> anyhow::Result<String> {
 
     Ok(format!("{:x}", hasher.finalize()))
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{fs::remove_file, io::Write};
-
-    use super::*;
-
-    #[test]
-    fn test_sha256checksum() {
-        let file_path = PathBuf::from("/tmp/test_checksum.txt");
-
-        let mut file = File::create(&file_path).unwrap();
-        file.write_all(b"hi!").unwrap();
-
-        let checksum = sha256checksum(&file_path).unwrap();
-
-        remove_file(&file_path).unwrap();
-
-        assert_eq!(
-            checksum,
-            "c0ddd62c7717180e7ffb8a15bb9674d3ec92592e0b7ac7d1d5289836b4553be2"
-        );
-    }
-}
