--- src/utils/default_true.rs
@@ -1,3 +1,0 @@
-pub const fn default_true() -> bool {
-    true
-}
